app/requirements.txt
dotenv==0.9.9
fastapi==0.115.11
supabase
pyjwt
python-on-whales
